chore: Divide and conquer
